# Conflicts:
#	pom.xml
#	src/main/java/com/mastercard/pts/integrated/issuing/configuration/TestConfiguration.java
#	src/main/java/com/mastercard/pts/integrated/issuing/pages/customer/cardmanagement/DeviceCardPackTemplatePage.java
#	src/main/java/com/mastercard/pts/integrated/issuing/pages/customer/processingcenter/InstitutionPage.java
#	src/main/java/com/mastercard/pts/integrated/issuing/steps/TestHooks.java
#	src/main/java/com/mastercard/pts/integrated/issuing/utils/ConstantData.java
#	src/main/java/com/mastercard/pts/integrated/issuing/utils/restapi/JsonUtil.java
#	src/main/java/com/mastercard/pts/integrated/issuing/utils/restapi/RestAssuredConfiguration.java
#	src/main/resources/config/demo/Data/TestData.xls
#	src/main/resources/config/demo/TestData/TestData.xlsx
#	src/main/resources/config/demo/apitest.properties
#	src/main/resources/webServiceRequestJsons/sample.json

+++ src/main/java/com/mastercard/pts/integrated/issuing/steps/TestHooks.java
@@ -45,30 +45,34 @@ public class TestHooks {
 	@BeforeStory
 	public void initStoryContext(@Named("StoryName") String storyName) {
 		testContext.initStoryContext(storyName);
-		if (storyName != "") {
-			Optional<Map<String, String>> data = dataLoader.loadData(storyName);
-			if (data.isPresent()) {
-				testContext.put(TestContext.KEY_STORY_DATA, data.get());
-			} else {
-				logger.info("There is no data set for story {}", storyName);
-			}
+		Optional<Map<String, String>> data = dataLoader.loadData(storyName);
+		if (data.isPresent()) {
+			testContext.put(TestContext.KEY_STORY_DATA, data.get());
+		} else {
+			logger.info("There is no data set for story {}", storyName);
 		}
-		logger.info("There is no data required for API  story");
-
 	}
 
 	@BeforeScenario
 	public void initTimeouts() {
-		Timeouts timeouts = driverProvider.get().manage().timeouts();
-		timeouts.implicitlyWait(imlicitWaitTimeout, TimeUnit.MILLISECONDS);
-		timeouts.pageLoadTimeout(pageWaitTimeout, TimeUnit.SECONDS);
-		timeouts.setScriptTimeout(scriptWaitTimeout, TimeUnit.SECONDS);
-		testContext.put("DRIVER", driverProvider.get());
+		if (System.getProperty("storyType") == null) {
+			Timeouts timeouts = driverProvider.get().manage().timeouts();
+			timeouts.implicitlyWait(imlicitWaitTimeout, TimeUnit.MILLISECONDS);
+			timeouts.pageLoadTimeout(pageWaitTimeout, TimeUnit.SECONDS);
+			timeouts.setScriptTimeout(scriptWaitTimeout, TimeUnit.SECONDS);
+			testContext.put("DRIVER", driverProvider.get());
+		} else {
+			logger.info("driverProvider not required for API  story");
+		}
 	}
 
 	// @AfterScenario
 	public void clearCookies() {
-		driverProvider.get().manage().deleteAllCookies();
+		if (System.getProperty("storyType") == null) {
+			driverProvider.get().manage().deleteAllCookies();
+		} else {
+			logger.info("driverProvider no required for API  story");
+		}
 	}
 
 	@AfterStory
